+++ public/vocables/index.html
@@ -17,7 +17,7 @@
         <script src="js/external/visjs/vis-timeline.min.js"></script>
         <link href="js/external/visjs/graph2d.min.css" rel="stylesheet" type="text/css" />
 
-        <script type="text/javascript" src="./js/external/visjs/vis-network.min.js"></script>
+        <script type="text/javascript" src="./js/external/visjs/vis-network.min.2.js"></script>
         <link href="js/external/visjs/vis-network.css" rel="stylesheet" type="text/css" />
         <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/jstree/3.2.1/themes/default/style.min.css" />
